+++ 1_SpringBoot/src/main/java/kh/springboot/admin/controller/AdminController.java
@@ -15,6 +15,8 @@ import org.springframework.web.bind.annotation.RequestMapping;
 
 import kh.springboot.board.model.service.BoardService;
 import kh.springboot.board.model.vo.Board;
+import kh.springboot.member.model.service.MemberService;
+import kh.springboot.member.model.vo.Member;
 import lombok.RequiredArgsConstructor;
 
 
@@ -24,6 +26,7 @@ import lombok.RequiredArgsConstructor;
 public class AdminController {
 	
 	private final BoardService bService;
+	private final MemberService mService;
 
 	@GetMapping("home")
 	public String moveToMainAdmin(Model model) {
@@ -76,4 +79,32 @@ public class AdminController {
 		
 		return "admin";
 	}
+	
+	@GetMapping("members")
+	public String selectMembers(Model model) {
+		ArrayList<Member> list = mService.selectMembers();
+		model.addAttribute("list", list);
+		return "members";
+	}
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
 }
