+++ 1_SpringBoot/src/main/java/kh/springboot/member/model/service/MemberService.java
@@ -86,4 +86,12 @@ public class MemberService {
 		return mapper.deleteTodo(num);
 	}
 
+	public ArrayList<Member> selectMembers() {
+		return mapper.selectMembers();
+	}
+
+	public int updateMemberItem(HashMap<String, String> map) {
+		return mapper.updateMemberItem(map);
+	}
+
 }
