+++ 1_SpringBoot/src/main/resources/templates/views/common/adminbar.html
@@ -30,14 +30,14 @@
 						</a>
 					</li>
 					<li class="nav-item">
-						<a class="nav-link d-flex align-items-center gap-2" href="members.adm"> 
+						<a class="nav-link d-flex align-items-center gap-2" href="/admin/members"> 
 							<svg class="bi">
 								<use xlink:href="#people" />
 							</svg> Members
 						</a>
 					</li>
 					<li class="nav-item">
-						<a class="nav-link d-flex align-items-center gap-2" href="boards.adm">
+						<a class="nav-link d-flex align-items-center gap-2" href="/admin/boards">
 							<svg class="bi">
 								<use xlink:href="#file-earmark-text" />
 							</svg> General Board
