+++ 1_SpringBoot/src/main/resources/templates/views/member/myinfo.html
@@ -285,7 +285,8 @@
 					const message = label.children[1].children[0].className.baseVal.includes('fill') ? '중요 표시된 일정입니다. ' : '';
 					if(confirm(message + '정말 삭제하시겠습니까?')){
 						$.ajax({
-							url: '/member/ldelete',
+							url: '/member/list',
+							type: 'delete',	// 삭제니까 delete
 							data: {num:todoNum},
 							success: data => {
 								//console.log(data);
@@ -311,7 +312,8 @@
 					todo.focus();
 				}else{
 					$.ajax({
-						url: '/member/linsert',
+						url: '/member/list',	// RestController위해서 주소 변경
+						type: 'post',			// RestController위해서 type 추가
 						data: {todo:todoValue, writer:loginUser.id},
 						success: data => {
 							//console.log(data);
@@ -360,7 +362,8 @@
 					const checked = this.checked ? 'C' : 'I';
 					
 					$.ajax({
-						url: '/member/lupdate',
+						url: '/member/list',
+						type: 'put',	//update라서 put
 						data: {status:checked, todoNum:this.value},
 						success: data => {
 							console.log(data);
@@ -392,7 +395,8 @@
 					const important = this.children[0].className.baseVal.indexOf('fill') > 0 ? 'N' : 'Y';
 					
 					$.ajax({
-						url: '/member/lupdate',
+						url: '/member/list',
+						type: 'put',
 						data: {important:important, todoNum:this.previousElementSibling.value},
 						success: data => {
 							//console.log(data);
@@ -433,7 +437,7 @@
 			$.ajax({
 				url: ' /member/profile',
 				data: formData,
-				type: 'post',
+				type: 'put',	//RestController 사용위해 post에서 put으로 변경
 				processData: false, // data를 String으로 변환 여부
 				contentType: false, // 전달하는 데이터 형식(false: 설정하지 않음, application/x-www-form-urlencoded)
 				success: data => {
